Move test_hadamard to test_operators.py
